+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 + core: Add `allow_multiple_instances()` to `RelmApp` to allow multiple concurrent instances of the same application
 + core: Add extension traits for `adw::PreferencesPage`
++ examples: drop sub-components instead of hiding
 
 ### Changed